pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmfr-data-viewer-backend"
version = "0.1.0"
description = "FastAPI backend for the TMFR telemetry data viewer."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pydantic>=2.0,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pyserial>=3.5",
  "pandas>=2.0",
]

[tool.setuptools]
packages = ["backend"]
